# swaggerui/swagger.yaml
basePath: /api/v1
consumes:
    - application/json
definitions:
    AggMachineResourceUsage:
        properties:
            bucket:
                type: string
                x-go-name: Bucket
            metric1:
                format: double
                type: number
                x-go-name: Metric1
            metric2:
                format: double
                type: number
                x-go-name: Metric2
            metric3:
                format: double
                type: number
                x-go-name: Metric3
        type: object
        x-go-package: multi-site-dashboard-go/internal/domain
    HTTPValidationError:
        properties:
            errors:
                items:
                    type: string
                type: array
                x-go-name: Errors
            message:
                type: string
                x-go-name: Message
        type: object
        x-go-package: multi-site-dashboard-go/internal/delivery/rest/handler
    MachineResourceUsage:
        properties:
            createdAt:
                type: string
                x-go-name: CreatedAt
            machine:
                type: string
                x-go-name: Machine
            metric1:
                format: int32
                type: integer
                x-go-name: Metric1
            metric2:
                format: int32
                type: integer
                x-go-name: Metric2
            metric3:
                format: int32
                type: integer
                x-go-name: Metric3
        type: object
        x-go-package: multi-site-dashboard-go/internal/domain
info:
    description: Documentation for Multi-Site Dashboard
    version: 1.0.0
paths:
    /machine:
        post:
            operationId: CreateMachineResourceUsage
            parameters:
                - in: body
                  name: machine
                  schema:
                    type: string
                  x-go-name: Machine
                - format: int32
                  in: query
                  name: metric1
                  type: integer
                  x-go-name: Metric1
                - format: int32
                  in: query
                  name: metric2
                  type: integer
                  x-go-name: Metric2
                - format: int32
                  in: query
                  name: metric3
                  type: integer
                  x-go-name: Metric3
            responses:
                "200":
                    description: MachineResourceUsage
                    schema:
                        $ref: '#/definitions/MachineResourceUsage'
                "400":
                    description: HTTPValidationError
                    schema:
                        $ref: '#/definitions/HTTPValidationError'
            tags:
                - ResourceTracking
    /machines/{machine}:
        get:
            operationId: GetAggMachineResourceUsage
            parameters:
                - in: path
                  name: machine
                  required: true
                  type: string
                  x-go-name: Machine
                - in: query
                  name: timeBucket
                  type: string
                  x-go-name: TimeBucket
                - in: query
                  name: lookBackPeriod
                  type: string
                  x-go-name: LookBackPeriod
            responses:
                "200":
                    description: AggMachineResourceUsage
                    schema:
                        items:
                            $ref: '#/definitions/AggMachineResourceUsage'
                        type: array
                "400":
                    description: HTTPValidationError
                    schema:
                        $ref: '#/definitions/HTTPValidationError'
            tags:
                - ResourceTracking
produces:
    - application/json
schemes:
    - http
    - https
swagger: "2.0"